+++ sparse_structures/DOKExcitationVector.hpp
@@ -30,7 +30,7 @@ template <typename T>
 class DOKExcitationVector : public IExcitationVector<T>
 {
 public:
-    DOKExcitationVector(const size_t num_rows);
+    DOKExcitationVector(const unsigned long num_rows);
     DOKExcitationVector(const IExcitationVector<T>& base);
 
     DOKExcitationVector(const DOKExcitationVector& rhs) = delete;
@@ -38,42 +38,42 @@ public:
     DOKExcitationVector(DOKExcitationVector&& rhs) = delete;
     DOKExcitationVector&& operator=(DOKExcitationVector&& rhs) = delete;
 
-    virtual void SetElement(const size_t r, const T& value);
-    virtual void SetElementQuick(const size_t r, const T& value);
+    virtual void SetElement(const unsigned long r, const T& value);
+    virtual void SetElementQuick(const unsigned long r, const T& value);
 
-    virtual T GetElement(const size_t r) const;
-    virtual T GetElementQuick(const size_t r) const;
+    virtual T GetElement(const unsigned long r) const;
+    virtual T GetElementQuick(const unsigned long r) const;
 
-    virtual size_t get_num_rows() const { return num_rows_; }
+    virtual unsigned long get_num_rows() const { return num_rows_; }
 
-    virtual size_t GetNnz() const { return map_.size(); }
+    virtual unsigned long GetNnz() const { return map_.size(); }
 
     virtual void Add(const IExcitationVector<T>& other);
-    virtual std::set<std::pair<size_t, T>> GetNzElements() const;
+    virtual std::set<std::pair<unsigned long, T>> GetNzElements() const;
 
 private:
-    const size_t num_rows_;
+    const unsigned long num_rows_;
 
-    std::unordered_map<size_t, T> map_;
+    std::unordered_map<unsigned long, T> map_;
 };
 
 template <typename T>
-DOKExcitationVector<T>::DOKExcitationVector(const size_t num_rows) : num_rows_(num_rows)
+DOKExcitationVector<T>::DOKExcitationVector(const unsigned long num_rows) : num_rows_(num_rows)
 {}
 
 template <typename T>
 DOKExcitationVector<T>::DOKExcitationVector(const IExcitationVector<T> &base) : num_rows_(base.get_num_rows())
 {
-    for (const std::pair<size_t, T>& element : base.GetNzElements())
+    for (const std::pair<unsigned long, T>& element : base.GetNzElements())
         map_[element.first] = element.second;
 }
 
 template <typename T>
-void DOKExcitationVector<T>::SetElement(const size_t r, const T& value)
+void DOKExcitationVector<T>::SetElement(const unsigned long r, const T& value)
 {
     IExcitationVector<T>::CheckBounds(r);
     if (IsNearlyEqual(value, 0.0)) {
-        typename std::unordered_map<size_t, T>::iterator it = map_.find(r);
+        typename std::unordered_map<unsigned long, T>::iterator it = map_.find(r);
         if (it != map_.end())
             map_.erase(it);
     } else
@@ -81,20 +81,20 @@ void DOKExcitationVector<T>::SetElement(const size_t r, const T& value)
 }
 
 template <typename T>
-void DOKExcitationVector<T>::SetElementQuick(const size_t r, const T &value)
+void DOKExcitationVector<T>::SetElementQuick(const unsigned long r, const T &value)
 {
     map_[r] = value;
 }
 
 template <typename T>
-T DOKExcitationVector<T>::GetElement(const size_t r) const
+T DOKExcitationVector<T>::GetElement(const unsigned long r) const
 {
     IExcitationVector<T>::CheckBounds(r);
     return GetElementQuick(r);
 }
 
 template <typename T>
-T DOKExcitationVector<T>::GetElementQuick(const size_t r) const
+T DOKExcitationVector<T>::GetElementQuick(const unsigned long r) const
 {
     T result;
 
@@ -109,16 +109,16 @@ T DOKExcitationVector<T>::GetElementQuick(const size_t r) const
 
 template <typename T>
 void DOKExcitationVector<T>::Add(const IExcitationVector<T>& other) {
-    for (const std::pair<size_t, T>& element : other.GetNzElements()) {
+    for (const std::pair<unsigned long, T>& element : other.GetNzElements()) {
         SetElement(element.first, PositiveClip(element.second + GetElement(element.first)));
     }
 }
 
 template <typename T>
-std::set<std::pair<size_t, T>> DOKExcitationVector<T>::GetNzElements() const
+std::set<std::pair<unsigned long, T>> DOKExcitationVector<T>::GetNzElements() const
 {
-    typename std::set<std::pair<size_t, T>> result;
-    for (typename std::unordered_map<size_t, T>::const_iterator it = map_.begin(); it != map_.end(); ++it)
+    typename std::set<std::pair<unsigned long, T>> result;
+    for (typename std::unordered_map<unsigned long, T>::const_iterator it = map_.begin(); it != map_.end(); ++it)
         result.insert(*it);
 
     return result;
