+++ sparse_structures/IExcitationVector.hpp
@@ -28,17 +28,17 @@ template <typename T>
 class IExcitationVector
 {
 public:
-    virtual void SetElement(const size_t r, const T& value) = 0;
-    virtual void SetElementQuick(const size_t r, const T& value) = 0;
+    virtual void SetElement(const unsigned long r, const T& value) = 0;
+    virtual void SetElementQuick(const unsigned long r, const T& value) = 0;
 
-    virtual T GetElement(const size_t r) const = 0;
-    virtual T GetElementQuick(const size_t r) const = 0;
+    virtual T GetElement(const unsigned long r) const = 0;
+    virtual T GetElementQuick(const unsigned long r) const = 0;
 
-    virtual size_t get_num_rows() const = 0;
+    virtual unsigned long get_num_rows() const = 0;
 
-    virtual size_t GetNnz() const = 0;
+    virtual unsigned long GetNnz() const = 0;
 
-    void CheckBounds(const size_t r) const
+    void CheckBounds(const unsigned long r) const
     {
         if (r >= get_num_rows())
             throw std::out_of_range("1D Out of Range");
@@ -46,14 +46,14 @@ public:
 
     virtual void Add(const IExcitationVector& other) = 0;
 
-    virtual std::set<std::pair<size_t, T>> GetNzElements() const = 0;
+    virtual std::set<std::pair<unsigned long, T>> GetNzElements() const = 0;
 
     std::string ToString() const
     {
-        const size_t num_rows = get_num_rows();
+        const unsigned long num_rows = get_num_rows();
 
         std::string str = std::string("ExcitationVector [num_lines=") + std::to_string(num_rows) + "]\n";
-        for (size_t r = 0; r < num_rows; ++r) {
+        for (unsigned long r = 0; r < num_rows; ++r) {
             str += std::to_string(GetElement(r)) + " ";
             str += "\n";
         }
